remove authen get items

+++ resources/item.py
@@ -1,6 +1,8 @@
-import sqlite3
 from flask_restful import Resource, reqparse
-from flask_jwt import jwt_required
+from flask_jwt_extended import (jwt_required, 
+    get_jwt, 
+    get_jwt_identity
+)
 from models.item import ItemModel
 
 class Item(Resource):
@@ -17,7 +19,7 @@ class Item(Resource):
     )
 
 
-    @jwt_required()
+    # @jwt_required()
     def get(self, name):
         item = ItemModel.find_item_by_name(name)
 
@@ -26,7 +28,7 @@ class Item(Resource):
 
         return {'message': 'Item not found'},404
 
-
+    @jwt_required(fresh=True)
     def post(self, name):
         data = Item.parser.parse_args()
         item = ItemModel(name,**data)
@@ -41,8 +43,11 @@ class Item(Resource):
 
         return item.json(),201
 
-
+    @jwt_required()
     def delete(self, name):
+        claims = get_jwt()
+        if not claims['is_admin']:
+            return {"message":"You dont have permission access this file"}
         item = ItemModel.find_item_by_name(name)
         if item:
             item.delete_to_db()
@@ -64,7 +69,19 @@ class Item(Resource):
 
 
 class ItemList(Resource):
+    # @jwt_required(optional=True)
+    # def get(self):
+    #     user_id = get_jwt_identity()
+    #     items = [item.json() for item in ItemModel.find_all()]
+    #     if user_id:
+    #         return {'items': items},200
+    #     return {
+    #         'items':[item['name'] for item in items],
+    #         'message': 'Data available when you log in'
+    #         },200
+    
     def get(self):
-        return {'items':[item.json() for item in ItemModel.query.all()]}
+        items = [item.json() for item in ItemModel.find_all()]
+        return {'items': items},200
